% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{normalize_family}
\alias{normalize_family}
\title{Normalize user input into a stats family object}
\usage{
normalize_family(family = NULL)
}
\arguments{
\item{family}{optional character scalar (e.g. "binomial"),
a function (e.g. \code{stats::binomial}), or a family object
(e.g. \code{stats::binomial()}).}
}
\value{
A stats family object
}
\description{
Normalize user input into a stats family object
}
