% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{validate_family_supported}
\alias{validate_family_supported}
\title{Validate the family/link combination is supported}
\usage{
validate_family_supported(fam)
}
\arguments{
\item{fam}{stats family object}
}
\value{
invisibly \code{TRUE}, errors otherwise
}
\description{
Validate the family/link combination is supported
}
