% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.capnet}
\alias{predict.capnet}
\title{Predict from a fitted \code{capnet} model}
\usage{
\method{predict}{capnet}(x, newx = NULL, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"capnet"}}

\item{newx}{Optional numeric matrix for prediction. If \code{NULL}, uses
\code{x$newx}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Numeric vector of predictions (length = number of rows in
\code{newx}).
}
\description{
Predict from a fitted \code{capnet} model
}
\seealso{
\code{\link[=capnet]{capnet()}}, \code{\link[=predict]{predict()}}
}
