% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.capnet}
\alias{predict.capnet}
\title{Predict from a fitted \code{capnet} model}
\usage{
\method{predict}{capnet}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{A fitted object of class \code{"capnet"}}

\item{newdata}{Optional numeric matrix for prediction. If \code{NULL}, uses
\code{object$newx}}

\item{type}{"link" returns linear predictor eta; "response" returns mean mu.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Numeric vector of predictions (length = number of rows in
\code{newx}).
}
\description{
Predict from a fitted \code{capnet} model
}
\seealso{
\code{\link[=capnet]{capnet()}}, \code{\link[=predict]{predict()}}
}
