% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.walk_capnet}
\alias{predict.walk_capnet}
\title{Predict from a walk-forward \code{capnet} fit}
\usage{
\method{predict}{walk_capnet}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"walk_capnet"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Numeric vector (or matrix) of predictions stored in x
}
\description{
Predict from a walk-forward \code{capnet} fit
}
\seealso{
\code{\link[=walk_capnet]{walk_capnet()}}, \code{\link[=predict]{predict()}}
}
