% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.capnet_violations}
\alias{print.capnet_violations}
\title{Print the violation matrix from \code{capnet_violations} object}
\usage{
\method{print}{capnet_violations}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"capnet_violations"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Sparse numeric matrix of contribution cap violations in fitted object x
}
\description{
Print the violation matrix from \code{capnet_violations} object
}
\seealso{
\code{\link[=capnet_violations]{capnet_violations()}}, \code{\link[=print]{print()}}
}
